Move common utilities to common directory.
