+++ locales/ru.json
@@ -71,5 +71,5 @@
   "error": {
     "back-home": "Back home",
     "not-found": "The page you are looking for was not found."
-  }  
+  }
 }
