+++ src/common/filters/forbidden-exception.filter.ts
