+++ src/common/filters/github-unathorized-exception.filter.ts
@@ -2,7 +2,7 @@ import { ExceptionFilter, Catch, ArgumentsHost, HttpStatus } from '@nestjs/commo
 import * as i18n from 'i18n';
 
 import { Request, Response } from 'express';
-import { GitHubUnauthorizedException } from '../../../utils/custom-errors';
+import { GitHubUnauthorizedException } from '../utils/custom-errors';
 
 @Catch(GitHubUnauthorizedException)
 export class GitHubUnauthorizedExceptionFilter implements ExceptionFilter {
