+++ src/common/guards/index.ts
@@ -1,3 +1,4 @@
 export * from './authenticated.guard';
 export * from './login.guard';
 export * from './github.guard';
+export * from './role.guard';
