+++ src/common/utils/youtube-preview.util.ts
